Updated controllers, and Added new view model
