+++ prime-dotnet-webapi/ViewModels/SiteRegistration/CreateOrganizationViewModel.cs
@@ -1,0 +1,7 @@
+namespace Prime.ViewModels
+{
+    public class CreateOrganizationViewModel
+    {
+        public int PartyId { get; set; }
+    }
+}
